requirements.txt
amqp==5.0.6
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.4b2
bsedata==0.4.0
celery==5.0.5
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
decorator==5.0.7
distlib==0.3.1
Django==3.2
django-celery-beat==2.2.0
django-extensions==3.1.3
django-filter==2.4.0
django-otp==1.0.4
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.1.2
filelock==3.0.12
flake8==3.9.1
gunicorn==20.1.0
identify==2.2.4
idna==2.10
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
lxml==4.6.3
mccabe==0.6.1
multitasking==0.0.9
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.12.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
twilio==6.57.0
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
yfinance==0.1.59
